Fast commit
