+++ WaifuGallery/Controls/Tabs/Preferences/Category.axaml
@@ -1,0 +1,32 @@
+﻿<Styles xmlns="https://github.com/avaloniaui"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:controls="using:WaifuGallery.Controls.Tabs.Preferences"
+        xmlns:ui="clr-namespace:FluentAvalonia.UI.Controls;assembly=FluentAvalonia">
+    <Design.PreviewWith>
+        <controls:Category Content="Test" Description="Test"  />
+    </Design.PreviewWith>
+
+    <Style Selector="controls|Category">
+        <!-- Set Defaults -->
+        <Setter Property="Template">
+            <ControlTemplate>
+                <Border CornerRadius="11" ClipToBounds="True" Margin="5">
+                    <ui:SettingsExpanderItem Description="{TemplateBinding Description}"
+                                             Content="{TemplateBinding Content}"
+                                             IconSource="{TemplateBinding IconSource}"
+                                             IsClickEnabled="True"
+                                             Command="{TemplateBinding Command}"
+                                             CommandParameter="{TemplateBinding CommandParameter}">
+                        <ui:SettingsExpanderItem.Footer>
+                            <ui:SymbolIcon Name="ExpandCollapseChevron"
+                                           Symbol="ChevronRight"
+                                           FontSize="18"
+                                           RenderTransform="none"
+                                           IsVisible="True" />
+                        </ui:SettingsExpanderItem.Footer>
+                    </ui:SettingsExpanderItem>
+                </Border>
+            </ControlTemplate>
+        </Setter>
+    </Style>
+</Styles>
