+++ WaifuGallery/Controls/Tabs/Preferences/FullScreen.axaml.cs
@@ -1,0 +1,13 @@
+﻿using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WaifuGallery.Controls.Tabs.Preferences;
+
+public partial class FullScreen : UserControl
+{
+    public FullScreen()
+    {
+        InitializeComponent();
+    }
+}
