+++ WaifuGallery/Controls/Tabs/Preferences/General.axaml
@@ -1,0 +1,23 @@
+﻿<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:controls="clr-namespace:FluentAvalonia.UI.Controls;assembly=FluentAvalonia"
+             xmlns:preferences="clr-namespace:WaifuGallery.ViewModels.Tabs.Preferences"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="WaifuGallery.Controls.Tabs.Preferences.General"
+             x:DataType="preferences:GeneralPreferencesViewModel"
+             IsVisible="{Binding IsVisible}">
+
+    <StackPanel>
+        <controls:SettingsExpander Header="Theme Variant"
+                                   IconSource="DarkTheme"
+                                   Description="Change the current theme variant">
+            <controls:SettingsExpander.Footer>
+                <ComboBox SelectedItem="{Binding CurrentThemeVariant}"
+                          ItemsSource="{Binding ThemesVariants}"
+                          MinWidth="150" />
+            </controls:SettingsExpander.Footer>
+        </controls:SettingsExpander>
+    </StackPanel>
+</UserControl>
