+++ WaifuGallery/Controls/Tabs/Preferences/StatusBar.axaml
@@ -1,0 +1,30 @@
+﻿<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:ui="clr-namespace:FluentAvalonia.UI.Controls;assembly=FluentAvalonia"
+             xmlns:preferences="clr-namespace:WaifuGallery.ViewModels.Tabs.Preferences"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="WaifuGallery.Controls.Tabs.Preferences.StatusBar"
+             x:DataType="preferences:StatusBarPreferencesViewModel"
+             IsVisible="{Binding IsVisible}">
+    <!-- <ui:SettingsExpander Header="Status Bar" -->
+    <!--                      Description="Auto Hide - Delay" -->
+    <!--                      IconSource="{DynamicResource BarricadeSvg}"> -->
+    <StackPanel>
+        <ui:SettingsExpander Header="Auto hide status bar"
+                             Description="idugh" IconSource="{DynamicResource EyeSvg}">
+            <ui:SettingsExpander.Footer>
+                <CheckBox IsChecked="{Binding AutoHideStatusBar}" />
+            </ui:SettingsExpander.Footer>
+        </ui:SettingsExpander>
+        <ui:SettingsExpander Header="Hide status bar delay"
+                                 Description="Min: 3000 Max:1000" IconSource="{DynamicResource TimerSvg}">
+            <ui:SettingsExpander.Footer>
+                <NumericUpDown Value="{Binding AutoHideStatusBarDelay}" Minimum="3000" Maximum="10000"
+                               Increment="500" IsEnabled="{Binding AutoHideStatusBar}" />
+            </ui:SettingsExpander.Footer>
+        </ui:SettingsExpander>
+    </StackPanel>
+    <!-- </ui:SettingsExpander> -->
+</UserControl>
