+++ WaifuGallery/Controls/Tabs/Preferences/StatusBar.axaml.cs
@@ -1,0 +1,13 @@
+﻿using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WaifuGallery.Controls.Tabs.Preferences;
+
+public partial class StatusBar : UserControl
+{
+    public StatusBar()
+    {
+        InitializeComponent();
+    }
+}
