+++ WaifuGallery/Controls/Tabs/Preferences/Tabs.axaml.cs
@@ -1,0 +1,13 @@
+﻿using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WaifuGallery.Controls.Tabs.Preferences;
+
+public partial class Tabs : UserControl
+{
+    public Tabs()
+    {
+        InitializeComponent();
+    }
+}
