+++ WaifuGallery/Controls/Tabs/PreferencesTab.axaml.cs
@@ -1,4 +1,7 @@
 ﻿using Avalonia.Controls;
+using Avalonia.Input;
+using Avalonia.Interactivity;
+using WaifuGallery.ViewModels.Tabs;
 
 namespace WaifuGallery.Controls.Tabs;
 
@@ -7,5 +10,13 @@ public partial class PreferencesTab : UserControl
     public PreferencesTab()
     {
         InitializeComponent();
+    }
+
+    private void InputElement_OnPointerPressed(object? sender, PointerPressedEventArgs e)
+    {
+        if (e.GetCurrentPoint(this).Properties.IsLeftButtonPressed)
+        {
+            (DataContext as PreferencesTabViewModel)?.GoToMainMenu();
+        }
     }
 }
