+++ WaifuGallery/Models/HotKeyManager.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Text.Json;
 using Avalonia.Input;
 using Serilog;
+using WaifuGallery.Converters;
 using WaifuGallery.Helpers;
 
 namespace WaifuGallery.Models;
