+++ WaifuGallery/Resources/Styles.axaml
@@ -1,5 +1,6 @@
 <Styles xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:ui="clr-namespace:FluentAvalonia.UI.Controls;assembly=FluentAvalonia">
     <Design.PreviewWith>
         <ThemeVariantScope RequestedThemeVariant="Dark">
             <Grid ColumnDefinitions="Auto,Auto,*,Auto,Auto,Auto" Background="{DynamicResource Brush.Background}"
@@ -33,6 +34,9 @@
             </Grid>
         </ThemeVariantScope>
     </Design.PreviewWith>
+    <Style Selector="ui|SettingsExpander">
+        <Setter Property="Margin" Value="5"></Setter>
+    </Style>
     <Style Selector="Svg">
         <Setter Property="Css" Value="{DynamicResource SvgCss}" />
     </Style>
