+++ WaifuGallery/ViewModels/Tabs/Preferences/GeneralPreferencesViewModel.cs
@@ -1,0 +1,48 @@
+﻿using System;
+using Avalonia;
+using Avalonia.Styling;
+using FluentAvalonia.Styling;
+using ReactiveUI;
+using WaifuGallery.Models;
+
+namespace WaifuGallery.ViewModels.Tabs.Preferences;
+
+public class GeneralPreferencesViewModel : ViewModelBase
+{
+    private bool _isVisible;
+    public string[] ThemesVariants { get; } = ["System", "Light", "Dark"];
+    private string _currentThemeVariant = string.Empty;
+
+    public GeneralPreferencesViewModel()
+    {
+        CurrentThemeVariant = Settings.Instance.Theme;
+        var fluentAvaloniaTheme = Application.Current?.Styles[0] as FluentAvaloniaTheme;
+        this.WhenAnyValue(x => x.CurrentThemeVariant)
+            .Subscribe(value =>
+            {
+                Settings.Instance.Theme = value;
+                var themeVariant = value switch
+                {
+                    "Light" => ThemeVariant.Light,
+                    "Dark" => ThemeVariant.Dark,
+                    _ => ThemeVariant.Default
+                };
+                if (Application.Current != null)
+                    Application.Current.RequestedThemeVariant = themeVariant;
+                if (fluentAvaloniaTheme != null)
+                    fluentAvaloniaTheme.PreferSystemTheme = value == "System";
+            });
+    }
+
+    public bool IsVisible
+    {
+        get => _isVisible;
+        set => this.RaiseAndSetIfChanged(ref _isVisible, value);
+    }
+
+    public string CurrentThemeVariant
+    {
+        get => _currentThemeVariant;
+        set => this.RaiseAndSetIfChanged(ref _currentThemeVariant, value);
+    }
+}
